Consolidate Electron IPC POC protocol in shared contracts
- Move the POC RPC group and methods to `@t3tools/contracts`
- Inline the renderer client setup in the IPC test and remove obsolete example files
- Update the main window placeholder copy
